requirements.txt
# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
websockets==12.0
python-multipart==0.0.9

# LLM
anthropic==0.28.0

# STT - Deepgram
deepgram-sdk==3.4.0

# TTS - ElevenLabs
elevenlabs==1.1.2

# RAG - Embeddings & Vector DB
openai==1.35.0               # for text-embedding-3-large
pinecone==7.3.0
langchain==0.2.5
langchain-community==0.2.5
langchain-openai==0.1.9

# Document parsing
pypdf==4.2.0
unstructured==0.14.6
python-docx==1.1.2

# Re-ranker
sentence-transformers==3.0.1

# BM25
rank-bm25==0.2.2

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.31
alembic==1.13.2

# Cache / Redis
redis[hiredis]==5.0.7

# HTTP client
aiohttp==3.9.5
httpx==0.27.0

# Config & Env
pydantic==2.7.4
pydantic-settings==2.3.4
python-dotenv==1.0.1

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.8

# Observability
opentelemetry-sdk==1.25.0
opentelemetry-api==1.25.0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-exporter-otlp==1.25.0
structlog==24.2.0
prometheus-fastapi-instrumentator==7.0.0

# Audio processing
numpy==1.26.4
soundfile==0.12.1
webrtcvad==2.0.10

# Utils
tenacity==8.4.1
python-ulid==2.2.0
orjson==3.10.5
